refatorado pagina de visao geral de turnos
