+++ apps/web/src/app/dashboard/turnos/types.ts
@@ -1,0 +1,16 @@
+export interface DadosGraficoTipoEquipe {
+  tipo: string;
+  quantidade: number;
+}
+
+export interface DadosGraficoHora {
+  hora: string;
+  tipo: string;
+  quantidade: number;
+}
+
+export interface DadosGraficoBase {
+  base: string;
+  tipo: string;
+  quantidade: number;
+}
